refactor: split worker base data read model
